BREAKING: Return a throttled function
This is a pretty heavy refactor that makes `async-throttle`
have the pretty API I've always dreamed that it should.

Previously you'd use throttle with `.map()` like so:

```js
Promise.all(urls.map((url) => throttle(async () => {
  // ...
})))
```

There's an anonymous function that is given the map argument, which
returns a call to `throttle()` which basically needs another anonymous
function to be passed to it in order to get the map() argument via
closure. Pretty nasty!

Now we can do this:

```js
const throttled = throttle(async url => {
  // ...
})
Promise.all(urls.map(throttled))
```

The `throttled` function is cached and the syntax is a lot cleaner.

The arguments passed when .map() is invoked get properly passed
to the throttled function, as well as the `this` context if that is set.

Tests updated and a few new cases added for the new behavior.